+++ src/controllers/DefaultController.php
@@ -14,6 +14,7 @@ use nystudio107\transcoder\Transcoder;
 
 use Craft;
 use craft\web\Controller;
+use craft\helpers\Json;
 
 /**
  * @author    nystudio107
@@ -31,7 +32,10 @@ class DefaultController extends Controller
      *         The actions must be in 'kebab-case'
      * @access protected
      */
-    protected $allowAnonymous = ['download-file'];
+    protected $allowAnonymous = [
+        'download-file',
+        'progress'
+    ];
 
     // Public Methods
     // =========================================================================
@@ -39,17 +43,87 @@ class DefaultController extends Controller
     /**
      * Force the download of a given $url.  We do it this way to prevent people
      * from downloading things that are outside of the server root.
+     *
+     * @param $url
      */
-    public function actionDownloadFile()
+    public function actionDownloadFile($url)
     {
-        $url = urldecode(Craft::$app->getRequest()->getParam('url'));
         $filePath = parse_url($url, PHP_URL_PATH);
-        $filePath = $_SERVER['DOCUMENT_ROOT'] . $filePath;
+        $filePath = $_SERVER['DOCUMENT_ROOT'].$filePath;
         Craft::$app->getResponse()->sendFile(
             $filePath,
             null,
             ['inline' => false]
         );
         Craft::$app->end();
+    }
+
+    /**
+     * Return a JSON-encoded array providing the progress of the transcoding:
+     *
+     * 'filename' => the name of the file
+     * 'duration' => the duration of the video/audio stream
+     * 'time' => the time of the current encoding
+     * 'progress' => a percentage indicating how much of the encoding is done
+     *
+     * @param $filename
+     *
+     * @return mixed
+     */
+    public function actionProgress($filename)
+    {
+        $result = [];
+        $progressFile = sys_get_temp_dir().DIRECTORY_SEPARATOR.$filename.".progress";
+        if (file_exists($progressFile)) {
+            $content = @file_get_contents($progressFile);
+            if ($content) {
+                // get duration of source
+                preg_match("/Duration: (.*?), start:/", $content, $matches);
+                $rawDuration = $matches[1];
+
+                // rawDuration is in 00:00:00.00 format. This converts it to seconds.
+                $ar = array_reverse(explode(":", $rawDuration));
+                $duration = floatval($ar[0]);
+                if (!empty($ar[1])) {
+                    $duration += intval($ar[1]) * 60;
+                }
+                if (!empty($ar[2])) {
+                    $duration += intval($ar[2]) * 60 * 60;
+                }
+
+                // Get the time in the file that is already encoded
+                preg_match_all("/time=(.*?) bitrate/", $content, $matches);
+                $rawTime = array_pop($matches);
+
+                // this is needed if there is more than one match
+                if (is_array($rawTime)) {
+                    $rawTime = array_pop($rawTime);
+                }
+
+                //rawTime is in 00:00:00.00 format. This converts it to seconds.
+                $ar = array_reverse(explode(":", $rawTime));
+                $time = floatval($ar[0]);
+                if (!empty($ar[1])) {
+                    $time += intval($ar[1]) * 60;
+                }
+                if (!empty($ar[2])) {
+                    $time += intval($ar[2]) * 60 * 60;
+                }
+
+                //calculate the progress
+                $progress = round(($time / $duration) * 100);
+
+                if ($progress < 100) {
+                    $result = [
+                        'filename' => $filename,
+                        'duration' => $duration,
+                        'time' => $time,
+                        'progress' => $progress,
+                    ];
+                }
+            }
+        }
+
+        return Json::encode($result);
     }
 }